Got the string escape to work on the double-quote
I just reverted all of the refactoring that I was using back to the more simple "String" definition. Works! I'm wondering if this is because if it is an "include" beneath the selector that handles the "begin" and "end" regexes, it takes a much lower precedence....but how does it work in all of the others?